+++ src/data/imbues/strange-and-unusual/spiked.ts
@@ -1,0 +1,265 @@
+import { tkey } from "../../../utils";
+import { RollString } from "../../../types";
+import { MaterialData } from "../../material";
+import { helpers } from "../../helpers";
+
+export function createImbueSpiked(): MaterialData {
+    const lkey = (
+        k: keyof Flatten<
+            Nested<
+                I18nKeyType,
+                "pf2e-monster-parts.data.imbuement.strange-and-unusual.spiked"
+            >
+        >,
+    ): I18nKey => tkey(`data.imbuement.strange-and-unusual.spiked.${k}`);
+
+    return {
+        type: "imbuement" as "imbuement",
+        key: "imbue:spiked",
+        itemPredicate: ["item:type:armor"],
+        monsterPredicate: [
+            { or: ["item:damage:type:piercing", "item:damage:type:slashing"] },
+            "item:type:melee",
+        ],
+    };
+    return [
+        {
+            ...base,
+            label: { type: "key", key: lkey("magic.label") },
+            header: {
+                description: { type: "key", key: lkey("flavor") },
+                labels: [
+                    ...helpers.leveledLabels(
+                        [10, 14, 18],
+                        ["1", "d4", "d6"],
+                        (damage: RollString) =>
+                            helpers.damage.label({
+                                type: "fire",
+                                value: damage,
+                            }),
+                    ),
+                    {
+                        levelMin: 2,
+                        text: {
+                            type: "key",
+                            key: "pf2e-monster-parts.data.imbuement.add-cantrip",
+                            parameters: {
+                                spell: "@UUID[Compendium.pf2e.spells-srd.Item.6DfLZBl8wKIV03Iq]",
+                            },
+                        },
+                        sort: 1,
+                    },
+                    ...helpers.leveledLabels(
+                        [4, 6, 8, 12, 16],
+                        [
+                            "magic.level-4-spells",
+                            "magic.level-6-spells",
+                            "magic.level-8-spells",
+                            "magic.level-12-spells",
+                            "magic.level-16-spells",
+                        ],
+                        (key: Parameters<typeof lkey>[0]) => ({
+                            text: { type: "key", key: lkey(key) },
+                            sort: 2,
+                        }),
+                    ),
+                    {
+                        levelMin: 20,
+                        text: {
+                            type: "key",
+                            key: lkey("magic.level-20-falling-stars"),
+                        },
+                        sort: 3,
+                    },
+                ],
+            },
+            effects: [
+                ...helpers.leveledEffects(
+                    [10, 14, 18],
+                    ["1", "d4", "d6"],
+                    (damage: RollString) =>
+                        helpers.damage.effect({
+                            type: "fire",
+                            value: damage,
+                            label: lkey("magic.label"),
+                        }),
+                ),
+            ],
+        },
+        {
+            ...base,
+            key: "imbue:fire:might",
+            label: { type: "key", key: lkey("might.label") },
+            header: {
+                description: { type: "key", key: lkey("flavor") },
+                labels: [
+                    ...helpers.leveledLabels(
+                        [4, 6, 8, 18],
+                        ["1", "d4", "d6", "d8"],
+                        (damage: RollString) =>
+                            helpers.damage.label({
+                                type: "fire",
+                                value: damage,
+                            }),
+                    ),
+                    {
+                        levelMin: 8,
+                        levelMax: 13,
+                        text: {
+                            type: "key",
+                            key: lkey("might.level-8-persistent"),
+                        },
+                        sort: 1,
+                    },
+                    {
+                        levelMin: 14,
+                        text: {
+                            type: "key",
+                            key: lkey("might.level-14-persistent"),
+                        },
+                        sort: 1,
+                    },
+                    {
+                        levelMin: 12,
+                        text: {
+                            type: "key",
+                            key: lkey("might.level-12-resistance"),
+                        },
+                        sort: 2,
+                    },
+                    {
+                        levelMin: 20,
+                        text: {
+                            type: "key",
+                            key: lkey("might.level-20-weakness"),
+                        },
+                        sort: 3,
+                    },
+                ],
+            },
+            effects: [
+                ...helpers.leveledEffects(
+                    [4, 6, 8, 18],
+                    ["1", "d4", "d6", "d8"],
+                    (damage: RollString) =>
+                        helpers.damage.effect({
+                            type: "fire",
+                            value: damage,
+                            label: lkey("might.label"),
+                        }),
+                ),
+                ...helpers.leveledEffects(
+                    [8, 14],
+                    ["1", "2"],
+                    (dice: RollString) => ({
+                        type: "RuleElement",
+                        rule: {
+                            key: "DamageDice",
+                            selector: "{item|_id}-damage",
+                            damageType: "fire",
+                            category: "persistent",
+                            dieSize: "d10",
+                            diceNumber: dice,
+                            label: lkey("might.label"),
+                            critical: true,
+                        },
+                    }),
+                ),
+            ],
+        },
+        {
+            ...base,
+            key: "imbue:fire:tech",
+            label: { type: "key", key: lkey("tech.label") },
+            header: {
+                description: { type: "key", key: lkey("flavor") },
+                labels: [
+                    {
+                        levelMin: 6,
+                        ...helpers.damage.label({
+                            type: "fire",
+                            value: 1,
+                        }),
+                    },
+                    ...helpers.leveledLabels(
+                        [4, 8, 14, 18],
+                        ["1", "d6", "d8", "d10"],
+                        (damage: RollString) =>
+                            helpers.damage.label({
+                                type: "fire",
+                                category: "persistent",
+                                value: damage,
+                            }),
+                    ),
+                    {
+                        levelMin: 8,
+                        text: {
+                            type: "key",
+                            key: lkey("tech.level-8-persistent"),
+                        },
+                        sort: 1,
+                    },
+                    {
+                        levelMin: 16,
+                        text: {
+                            type: "key",
+                            key: lkey("tech.level-16-off-guard"),
+                        },
+                        sort: 2,
+                    },
+                    {
+                        levelMin: 12,
+                        text: {
+                            type: "key",
+                            key: lkey("tech.level-12-resistance"),
+                        },
+                        sort: 3,
+                    },
+                    {
+                        levelMin: 20,
+                        text: {
+                            type: "key",
+                            key: lkey("tech.level-20-weakness"),
+                        },
+                        sort: 4,
+                    },
+                ],
+            },
+            effects: [
+                {
+                    levelMin: 6,
+                    ...helpers.damage.effect({
+                        type: "fire",
+                        value: 1,
+                        label: lkey("tech.label"),
+                    }),
+                },
+                ...helpers.leveledEffects(
+                    [4, 8, 14, 18],
+                    ["1", "d6", "d8", "d10"],
+                    (damage: RollString) =>
+                        helpers.damage.effect({
+                            type: "fire",
+                            category: "persistent",
+                            value: damage,
+                            label: lkey("tech.label"),
+                        }),
+                ),
+                {
+                    levelMin: 8,
+                    type: "RuleElement",
+                    rule: {
+                        key: "DamageDice",
+                        selector: "{item|_id}-damage",
+                        damageType: "fire",
+                        category: "persistent",
+                        dieSize: "d10",
+                        diceNumber: 1,
+                        label: lkey("tech.label"),
+                        critical: true,
+                    },
+                },
+            ],
+        },
+    ];
+}